Initial refactoring of "character control"; abstract what is now "character control" into a general "base control" from which "character control" can sub-class.
