+++ nucypher/blockchain/eth/interfaces.py
@@ -60,7 +60,7 @@ from nucypher.blockchain.eth.sol.compile.compile import multiversion_compile
 from nucypher.blockchain.eth.sol.compile.constants import SOLIDITY_SOURCE_ROOT
 from nucypher.blockchain.eth.sol.compile.types import SourceBundle
 from nucypher.blockchain.eth.utils import get_transaction_name, prettify_eth_amount
-from nucypher.characters.control.emitters import JSONRPCStdoutEmitter, StdoutEmitter
+from nucypher.control.emitters import StdoutEmitter, JSONRPCStdoutEmitter
 from nucypher.utilities.ethereum import encode_constructor_arguments
 from nucypher.utilities.gas_strategies import (
     construct_datafeed_median_strategy,
