+++ nucypher/blockchain/eth/signers/hardware.py
@@ -33,7 +33,7 @@ from web3 import Web3
 
 from nucypher.blockchain.eth.decorators import validate_checksum_address
 from nucypher.blockchain.eth.signers.base import Signer
-from nucypher.characters.control.emitters import StdoutEmitter
+from nucypher.control.emitters import StdoutEmitter
 
 
 def handle_trezor_call(device_func):
