+++ nucypher/characters/control/specifications/fields/cleartext.py
@@ -19,7 +19,7 @@ from base64 import b64encode
 
 from marshmallow import fields
 
-from nucypher.characters.control.specifications.fields.base import BaseField
+from nucypher.control.specifications.fields.base import BaseField
 
 
 class Cleartext(BaseField, fields.String):
