+++ nucypher/characters/control/specifications/fields/key.py
@@ -18,8 +18,9 @@
 from marshmallow import fields
 from umbral.keys import UmbralPublicKey
 
-from nucypher.characters.control.specifications.exceptions import InvalidInputData, InvalidNativeDataTypes
-from nucypher.characters.control.specifications.fields.base import BaseField
+from nucypher.characters.control.specifications.exceptions import InvalidNativeDataTypes
+from nucypher.control.specifications.exceptions import InvalidInputData
+from nucypher.control.specifications.fields.base import BaseField
 
 
 class Key(BaseField, fields.Field):
