+++ nucypher/characters/control/specifications/fields/misc.py
@@ -15,31 +15,8 @@
  along with nucypher.  If not, see <https://www.gnu.org/licenses/>.
 """
 
-import click
-from marshmallow import fields
-
-from nucypher.characters.control.specifications.exceptions import InvalidInputData
-from nucypher.characters.control.specifications.fields.base import BaseField
+from nucypher.control.specifications.fields.base import Integer, PositiveInteger
 from nucypher.cli import types
-
-
-class String(BaseField, fields.String):
-    pass
-
-
-class List(BaseField, fields.List):
-    pass
-
-
-class Integer(BaseField, fields.Integer):
-    click_type = click.INT
-
-
-class PositiveInteger(Integer):
-
-    def _validate(self, value):
-        if not value > 0:
-            raise InvalidInputData(f"{self.name} must be a positive integer.")
 
 
 class M(PositiveInteger):
@@ -52,9 +29,3 @@ class N(PositiveInteger):
 
 class Wei(Integer):
     click_type = types.WEI
-
-
-class click:
-    def __init__(self, *args, **kwargs):
-        self.args = args
-        self.kwargs = kwargs
