+++ nucypher/cli/actions/select.py
@@ -31,7 +31,7 @@ from nucypher.blockchain.eth.networks import NetworksInventory
 from nucypher.blockchain.eth.registry import InMemoryContractRegistry, BaseContractRegistry
 from nucypher.blockchain.eth.signers.base import Signer
 from nucypher.blockchain.eth.token import NU, Stake
-from nucypher.characters.control.emitters import StdoutEmitter
+from nucypher.control.emitters import StdoutEmitter
 from nucypher.cli.actions.configure import get_config_filepaths
 from nucypher.cli.literature import (
     GENERIC_SELECT_ACCOUNT,
