+++ nucypher/control/specifications/__init__.py
@@ -14,18 +14,3 @@
  You should have received a copy of the GNU Affero General Public License
  along with nucypher.  If not, see <https://www.gnu.org/licenses/>.
 """
-
-from marshmallow import fields
-
-from nucypher.control.specifications.fields.base import BaseField
-
-
-class Label(BaseField, fields.Field):
-
-    def _serialize(self, value, attr, obj, **kwargs):
-        return value.decode('utf-8')
-
-    def _deserialize(self, value, attr, data, **kwargs):
-        if isinstance(value, bytes):
-            return value
-        return value.encode()
