+++ tests/integration/characters/test_specifications.py
@@ -23,9 +23,8 @@ import pytest
 
 from nucypher.characters.control.specifications import fields
 from nucypher.characters.control.specifications.alice import GrantPolicy
-from nucypher.characters.control.specifications.base import BaseSchema
-from nucypher.characters.control.specifications.exceptions import (InvalidArgumentCombo, InvalidInputData,
-                                                                   SpecificationError)
+from nucypher.control.specifications.base import BaseSchema
+from nucypher.control.specifications.exceptions import SpecificationError, InvalidInputData, InvalidArgumentCombo
 from nucypher.crypto.powers import DecryptingPower
